requirements.txt
# Amazon Product Recommendation System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Spark/Hadoop
pyspark>=3.3.0
pyyaml>=5.4

# Data Processing
numpy>=1.20.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter (for notebook support)
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
scikit-learn>=0.24.0
scipy>=1.7.0

# Optional: Performance improvements
pyarrow>=5.0.0
